5.6.9-B2641.94
- Fix: POP3 Return -ERR instead of +ERR when inbox message fetch fails, see: [this](https://github.com/cybercode3/hmailserver/commit/bc31f5eedd2d39ac320caf9870e5191eb835067c) commit
- Fix: Prevent rewrite of message untouched message parts, see: [pull 506](https://github.com/hmailserver/hmailserver/pull/506)
- Fix: Stop SMTPDeliveryManager gracefully, see: [pull 507](https://github.com/hmailserver/hmailserver/pull/507)
- Fix: Fix false positives in line ending check, see: [pull 508](https://github.com/hmailserver/hmailserver/pull/508)
- Fix: Improved regex for validation of email addresses & domains, see: [pull 509](https://github.com/hmailserver/hmailserver/pull/509)
- Fix: Scheduled Expiry of Auto-reply doesnt clear the 'already replied' cache, see: [pull 510](https://github.com/hmailserver/hmailserver/pull/510)

+++ hmailserver/source/Server/Common/TCPIP/TCPConnection.h
@@ -74,7 +74,7 @@ namespace HM
 
       ConnectionState GetConnectionState() { return connection_state_;  }
 
-      int GetBufferSize() {return BufferSize; }
+      size_t GetBufferSize() {return BufferSize; }
 
       void SetTimeout(int seconds);
       AnsiString GetIPAddressString();
